refactor out repos route, update cvr
