+++ package.json
@@ -9,7 +9,7 @@
     "body-parser": "~1.12.0",
     "connect-flash": "^0.1.1",
     "cookie-parser": "~1.3.4",
-    "cvr": "^1.0.0",
+    "cvr": "^1.0.1",
     "debug": "~2.1.1",
     "express": "~4.12.2",
     "express-session": "^1.11.1",
